+++ vulkan-tutorial.vcxproj
@@ -156,10 +156,12 @@
     <ClCompile Include="CommandPoolManager.cpp" />
     <ClCompile Include="FileCaching.cpp" />
     <ClCompile Include="ImageLibraryImplementations.cpp" />
+    <ClCompile Include="LayoutsBuilder.cpp" />
     <ClCompile Include="Memory.cpp" />
     <ClCompile Include="MeshData.cpp" />
     <ClCompile Include="MeshLibraryImplementations.cpp" />
     <ClCompile Include="mikktspace.c" />
+    <ClCompile Include="PipelineDataObject.cpp"/>
     <ClCompile Include="SceneObjectData.cpp" />
     <ClCompile Include="ShaderLoading.cpp" />
     <ClCompile Include="textureCreation.cpp" />
@@ -174,6 +176,7 @@
     <ClInclude Include="common-structs.h" />
     <ClInclude Include="FileCaching.h" />
     <ClInclude Include="ImageLibraryImplementations.h" />
+    <ClInclude Include="LayoutsBuilder.h" />
     <ClInclude Include="Memory.h" />
     <ClInclude Include="MeshLibraryImplementations.h" />
     <ClInclude Include="textureCreation.h" />
